backend/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
youtube-transcript-api
python-dotenv
langchain-google-genai
langchain-community
pydantic
requests
